+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - Custom encoders can now receive `_parent` and `_sort_keys` parameters to enable proper encoding of nested structures. ([#429](https://github.com/python-poetry/tomlkit/issues/429))
 
+## Fixed
+
+- Fixed `tomlkit.boolean()` API to correctly handle boolean inputs. ([#442](https://github.com/python-poetry/tomlkit/issues/442))
+
 ## [0.13.3] - 2025-06-05
 
 ### Added