+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log
 
-## [Unreleased]
+## [0.14.0] - 2026-01-13
 
 ### Changed
 
@@ -452,7 +452,8 @@
 - Fixed handling of super tables with different sections.
 - Fixed raw strings escaping.
 
-[unreleased]: https://github.com/sdispater/tomlkit/compare/0.13.3...master
+[unreleased]: https://github.com/sdispater/tomlkit/compare/0.14.0...master
+[0.14.0]: https://github.com/sdispater/tomlkit/releases/tag/0.14.0
 [0.13.3]: https://github.com/sdispater/tomlkit/releases/tag/0.13.3
 [0.13.2]: https://github.com/sdispater/tomlkit/releases/tag/0.13.2
 [0.13.1]: https://github.com/sdispater/tomlkit/releases/tag/0.13.1